+++ demo_tablet_client/run_demo.sh
@@ -4,5 +4,5 @@
 # It's to allow a demo to be run whilst the main copy is broken for development!
 
 THISDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-export LD_LIBRARY_PATH=/home/rudolf/dev/qt_local_build/openssl_linux_x86_64_build/openssl-1.1.0g
+export LD_LIBRARY_PATH=/home/rudolf/dev/qt_local_build/openssl_linux_x86_64_build/openssl-1.1.1c
 $THISDIR/camcops